requirements.txt
azure-core==1.38.0
azure-identity==1.25.1
blinker==1.9.0
boto3==1.40.48
botocore==1.40.48
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.2.1
connect==0.2
cryptography==46.0.4
Flask==3.1.1
gunicorn==25.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
msal==1.34.0
msal-extensions==1.3.1
mssql==1.0.1
mssql-python==1.3.0
numpy==2.3.3
pandas==2.3.2
protobuf==6.32.1
pycparser==3.0
PyJWT==2.11.0
pymssql==2.3.11
pyodbc==5.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
